REFACTOR: Moved `CodeComponent` to `BlogViewer` folder
