+++ components/BlogViewer/CodeComponent/CodeComponentManager/CodeComponentManager.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren, useRef, useState } from "react";
-import { stringNumberContext } from "../../../../../context/keyValue";
+import { stringNumberContext } from "../../../../context/keyValue";
 
 interface Props {}
